:bug: Fix root user

+++ tests/test-composer.sh
@@ -9,6 +9,22 @@ if [ -z ${IMAGE_TAG} ]; then
     exit 1
 fi
 
+if [ `whoami` = root ]; then
+    echo "Running as root, change to another user"
+
+    if ! egrep -i "1000" /etc/group ; then
+        groupadd -f -g 1000 bob
+    fi
+
+    if ! id 1000 >/dev/null 2>/dev/null ; then
+        useradd -u 1000 -g bob bob && \
+        mkdir -p /home/bob && \
+        chown -R bob:bob .
+    fi
+    sudo -u 1000:1000 -E sh "./test-composer.sh ${IMAGE_TAG}"
+    exit $?
+fi
+
 if [ -d vendor ]; then
     rm -Rf vendor
 fi
